constants carousel was created and copypaste of utils-constants
