+++ src/components/Carousel/Carousel.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import Carousel from "react-material-ui-carousel";
-import PlayArrowIcon from '@material-ui/icons/PlayArrow';
+import PlayArrowIcon from "@material-ui/icons/PlayArrow";
 import { Paper } from "@material-ui/core";
-import URL_IMAGES  from "../../utils/constants";
+import URL_IMAGES  from "./constants";
 import useStyles from "./styles";
 
 function Item(props) {
@@ -24,7 +24,7 @@ function CustomCarousel() {
         PrevIcon={<PlayArrowIcon aria-label="imagen anterior" style={{transform: "rotate(-180deg)"}}/>}
         navButtonsProps={{
           style: {
-              backgroundColor: 'rgba(203,169,205,0.3)',
+              backgroundColor: "rgba(203,169,205,0.3)",
               borderRadius: 30,
               width: 30,
               height: 30,
