+++ src/utils/constants.js
@@ -4,19 +4,7 @@ export const orderRoute = '/pedido';
 
 export const imagebanner = 'https://firebasestorage.googleapis.com/v0/b/la-store-5dcee.appspot.com/o/bannershop.png?alt=media&token=fcd0aabf-ee64-4ce6-ba50-96e84409d49c';
 
-const URL_IMAGES = [
-  {
-    url: "https://firebasestorage.googleapis.com/v0/b/la-store-5dcee.appspot.com/o/carrousel%2FbannerCursosCarrousel_opt-min.png?alt=media&token=2bc8859f-3154-4b63-8dce-40f9f9c28424",
-  },
-  {
-    url: "https://firebasestorage.googleapis.com/v0/b/la-store-5dcee.appspot.com/o/carrousel%2FbannerDescuentos-min.png?alt=media&token=88c67a6d-dbb4-42d9-96d3-9aca77ec2a25",
-  },
-  {
-    url: "https://firebasestorage.googleapis.com/v0/b/la-store-5dcee.appspot.com/o/carrousel%2Fbanneroraculo_opt-min.png?alt=media&token=d3fc2b37-0450-4f2b-a5cc-ffc9dab65c88",
-  },
-];
-
-export const featuredProductsMockup = [
+const featuredProductsMockup = [
   {
     id: "1",
     name: "huevos yoni",
@@ -88,4 +76,4 @@ export const URL_SOCIAL = {
 export const taxes = [1.19];
 
 
-export default URL_IMAGES;
+export default featuredProductsMockup;
